Added wont fix to api service response for GET /images/{imageDigest}/vuln/{vtype}
