+++ anchore_engine/services/apiext/swagger/swagger.yaml
@@ -745,6 +745,7 @@ paths:
           in: query
           type: boolean
           required: false
+          description: Filter results to include only vulnerabilities that are not marked as invalid by upstream OS vendor data. When set to true, it will filter out all vulnerabilities where `wont_fix` is False. If false all vulnerabilities are returned regardless of `wont_fix`
         - $ref: "#/parameters/AsAccountParameter"
       responses:
         200:
@@ -3936,6 +3937,9 @@ definitions:
       package_path:
         type: string
         description: The location (if applicable) of the vulnerable package in the container filesystem
+      wont_fix:
+        type: string
+        description: Whether a vendor will fix or not fix the vulnerability
       nvd_data:
         $ref: "#/definitions/NvdDataList"
       vendor_data:
